Move loggerForRequest() into a separate package
loggerForRequest() is used across a bunch of different places.  Moving
it into a sub-package makes it easier to restructure other pieces into
sub-packages.
