+++ logger/logger.go
@@ -1,0 +1,24 @@
+package logger
+
+import (
+	"net/http"
+	"strings"
+
+	log "github.com/sirupsen/logrus"
+)
+
+func ForRequest(r *http.Request) *log.Entry {
+	return log.WithContext(r.Context()).WithFields(log.Fields{
+		"ip":      getUserIP(r),
+		"request": r.URL.String(),
+	})
+}
+
+func getUserIP(r *http.Request) string {
+	headerIP := r.Header.Get("X-Forwarded-For")
+	if headerIP != "" {
+		return headerIP
+	}
+
+	return strings.Split(r.RemoteAddr, ":")[0]
+}
